requirements.txt
cohere
gradio
faiss-cpu
langchain
langchain-cohere
langchain-community
python-dotenv
unstructured[all-docs]
tomli